requirements.txt
aiogram>=3.4.1
aiohttp>=3.9.3
yt-dlp>=2024.3.10
python-dotenv>=1.0.1
aiosqlite>=0.20.0
